App/backend/rest-api/requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pipflask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
flask-restful-swagger-2==0.35
flask-swagger-ui==4.11.1
yfinance===0.2.40